Move computeOldest to a package object
This method is going to be used in both CommandRunner and StatusChecker classes
